@@ -13,7 +13,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from typing import List, Optional
+from typing import Optional
 
 import numpy as np
 
@@ -21,8 +21,13 @@ from qustop.core.state import State
 
 
 class Ensemble:
-    def __init__(self, states: List[State], probs: Optional[List[float]] = None) -> None:
+    """A set of `State` objects denoting quantum states where each element has an associated
+    probability of being selected from the set.
+    """
+
+    def __init__(self, states: list[State], probs: Optional[list[float]] = None) -> None:
         """
+        Initializes an Ensemble
         If no probability vector is specified in `probs`, the default assumes
         a uniform probability distribution for all states in the ensemble.
         """
@@ -35,7 +40,7 @@ class Ensemble:
     def __str__(self) -> str:
         out_s = f"Ensemble: num_states = {len(self)}\n"
         for i, subsystem in enumerate(self._states):
-            if i == len(self._states)-1:
+            if i == len(self._states) - 1:
                 out_s += f"ρ_{subsystem}"
             else:
                 out_s += f"ρ_{subsystem} ⊗ "
@@ -47,30 +52,30 @@ class Ensemble:
     def __getitem__(self, key: int) -> State:
         return self.states[key]
 
-    def swap(self, sub_sys_swap: List[int]):
+    def swap(self, sub_sys_swap: list[int]):
         for state in self._states:
             state.swap(sub_sys_swap)
 
     @property
-    def probs(self) -> List[float]:
+    def probs(self) -> list[float]:
         return self._probs
 
     @property
-    def states(self) -> List[State]:
+    def states(self) -> list[State]:
         return self._states
 
     @property
-    def density_matrices(self) -> List[np.ndarray]:
+    def density_matrices(self) -> list[np.ndarray]:
         return [state.value for state in self._states]
 
     @staticmethod
-    def _prepare_states(states: List[State]) -> Optional[List[State]]:
+    def _prepare_states(states: list[State]) -> Optional[list[State]]:
         # Assume at least one state is provided.
         if states is None or states == []:
             raise ValueError("InvalidStates: There must be at least one state provided.")
         return states
 
-    def _prepare_probs(self, probs: List[float]) -> Optional[List[float]]:
+    def _prepare_probs(self, probs: list[float]) -> Optional[list[float]]:
         # If probability vector is not explicitly provided, assume the ensemble
         # has a uniform distribution.
         if probs is None:
@@ -79,8 +84,10 @@ class Ensemble:
         # The probability vector must be of the same length of the number of
         # states in the ensemble.
         if len(probs) != len(self):
-            raise ValueError("The number of probabilities in vector must be the same as the number "
-                             "of states in the ensemble.")
+            raise ValueError(
+                "The number of probabilities in vector must be the same as the number "
+                "of states in the ensemble."
+            )
 
         # Probability vector must sum to one to be valid.
         if not np.isclose(sum(probs), 1):
