@@ -13,5 +13,16 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from qustop._about import about
-from qustop._version import __version__
+import numpy as np
+
+from toqito.states import basis
+from qustop.core import State
+
+e_0, e_1 = basis(2, 0), basis(2, 1)
+
+
+def test_state_shape():
+    bell_vec = 1/np.sqrt(2) * (np.kron(e_0, e_0) + np.kron(e_1, e_1))
+    state_vec = State(bell_vec, [2, 2])
+
+    assert state_vec.shape == (4, 4)
