@@ -14,14 +14,13 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from typing import List
-
 import cvxpy
 import numpy as np
 
 
 class Positive:
     """Positive (global) distinguishability."""
+
     def __init__(self, ensemble, dist_method, fast=False):
         self.ensemble = ensemble
         self.states = self.ensemble.density_matrices
@@ -48,30 +47,33 @@ class Positive:
 
         # There is a closed-form expression for the distinguishability of two density matrices.
         if len(self.ensemble) == 2:
-            opt_val = 1/2 + np.linalg.norm(self.probs[0] * self.states[0] -
-                                           self.probs[1] * self.states[1])/2
+            opt_val = (
+                1 / 2
+                + np.linalg.norm(self.probs[0] * self.states[0] - self.probs[1] * self.states[1])
+                / 2
+            )
             D, V = np.linalg.eig(
-                self.probs[0] * self.states[0][:, [0]] @ self.states[0][:, [0]].conj().T -
-                self.probs[1] * self.states[1][:, [0]] @ self.states[1][:, [0]].conj().T
+                self.probs[0] * self.states[0][:, [0]] @ self.states[0][:, [0]].conj().T
+                - self.probs[1] * self.states[1][:, [0]] @ self.states[1][:, [0]].conj().T
             )
             D = np.diag(D)
             pind = np.argwhere(np.asarray(D) >= 0)
             print(V[:, [pind]] @ V[:, [pind]].conj().T)
             # pind = (D >= 0).nonzero()
-#            print(V[:, pind])
-#            meas_1 = V[:, pind] @ V[:, pind].conj().T
-#            print(V[:, pind])
+            #            print(V[:, pind])
+            #            meas_1 = V[:, pind] @ V[:, pind].conj().T
+            #            print(V[:, pind])
 
             # Construct optimal measurements:
             return opt_val, []
 
-        # # If just the optimal value is required, it is often less
-        # # computationally intensive to solve the dual problem.
-        # if self.fast:
-        #     return self.dual_problem()
-        # # Otherwise, return the optimal value and the optimal measurements for
-        # # obtaining that value.
-        # return self.primal_problem()
+        # If just the optimal value is required, it is often less
+        # computationally intensive to solve the dual problem.
+        if self.fast:
+            return self.dual_problem()
+        # Otherwise, return the optimal value and the optimal measurements for
+        # obtaining that value.
+        return self.primal_problem()
 
     def primal_problem(self):
         r"""
@@ -93,7 +95,9 @@ class Positive:
             for i, _ in enumerate(self.states):
                 for j, _ in enumerate(self.states):
                     if i != j:
-                        constraints.append(cvxpy.trace(self.states[i].conj().T @ measurements[i]) == 0)
+                        constraints.append(
+                            cvxpy.trace(self.states[i].conj().T @ measurements[i]) == 0
+                        )
 
         # Note we have one additional measurement operator in the unambiguous case.
         for i in range(num_measurements):
